Reduce price of field artillery to $4 to align with AA.
From Zach: Artillery: I might bring down the cost of the artillery to 4. Artillery cost 4 in Axis and Allies and I think that's a good balance. The problem is that if I'm spending 10 resources it's much more effective to purchase five infantry than two artillery. The infantry will win hands down. So unless you are capped at your production like Russia, it's not worth the cost to build the artillery.