Only allow a maximum of 6 harbor defenses in a stack
Zach: "One other thought, but do you think having a stack of eight coastal defenses is too many? With troops there are greater numbers so it isn't as formidable to topple a stack of trenches but ships are so much more expensive that taking out a stack of eight coastal defenses behind a fleet means you need a much much larger fleet and that is too expensive and costly. So what happens is that we keep having stalemates. German and British fleets. Russian and Ottoman fleets. Italian and Mediterranean fleets. British Commonwealth and German Afrika fleets. It's too expensive and too big to take out a full stack of coastal defenses that we end up just sitting behind them for the whole game. I wonder what would happen if you capped the limit of coastal defenses at 4?"